farmtech_project/requirements.txt
Django==5.1.7
django-crispy-forms==2.1.0
crispy-bootstrap5==0.7.0
mysqlclient==2.2.0
SQLAlchemy==2.0.23
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
scikit-learn==1.3.2
scipy==1.11.4
Pillow==10.1.0
opencv-python==4.8.1.78
google-generativeai==0.3.2
requests==2.31.0
urllib3==2.1.0
django-allauth==0.61.0
django-otp==1.3.0
argon2-cffi==23.1.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0
faker==20.1.0
python-dotenv==1.0.0
pytz==2023.3
tqdm==4.66.1